rename stores
