+++ packages/app/src/store/index.ts
@@ -1,14 +1,16 @@
 import { createPinia as _createPinia } from 'pinia'
 
 // Pinia client-side storage
-export * from './main'
+export * from './main-store'
 
 export * from './modals'
 
 export * from './specs-store'
 
+export * from './aut-store'
+
 // Mobx Store Wrapper from @packages/runner-shared
-export * from './runner-store'
+export * from './mobx-runner-store'
 
 // Reusable installation function, used as an entry point for tests that
 // require an identical setup to main.ts
